requirements.txt
alpaca-py==0.28.0
gemini_toolbox==0.5.0
python-dotenv==1.0.1
python-telegram-bot==20.6
python-dateutil==2.9.0.post0
